Initial version
This package is an abstraction on top of the go-tpm libraries to use a
local TPM to create and use RSA keys that are bound to that TPM. The
private keys can never be used without the TPM that was used to create
them.

Any number of keys can be created and used concurrently. The library
takes care loading the right key in the TPM, as needed.

All keys are 2048-bit RSA and non-duplicable.